optimisation and pep8 fixes

+++ cart/cart.py
@@ -1,22 +1,25 @@
 import datetime
-
+from django.db.models import Sum
+from django.db.models import F
 from . import models
 
 CART_ID = 'CART-ID'
 
+
 class ItemAlreadyExists(Exception):
     pass
 
+
 class ItemDoesNotExist(Exception):
     pass
+
 
 class Cart:
     def __init__(self, request):
         cart_id = request.session.get(CART_ID)
         if cart_id:
-            try:
-                cart = models.Cart.objects.get(id=cart_id, checked_out=False)
-            except models.Cart.DoesNotExist:
+            cart = models.Cart.objects.filter(id=cart_id, checked_out=False).first()
+            if cart is None:
                 cart = self.new(request)
         else:
             cart = self.new(request)
@@ -27,71 +30,46 @@ class Cart:
             yield item
 
     def new(self, request):
-        cart = models.Cart(creation_date=datetime.datetime.now())
-        cart.save()
+        cart = models.Cart.objects.create(creation_date=datetime.datetime.now())
         request.session[CART_ID] = cart.id
         return cart
 
     def add(self, product, unit_price, quantity=1):
-        try:
-            item = models.Item.objects.get(
-                cart=self.cart,
-                product=product,
-            )
-        except models.Item.DoesNotExist:
-            item = models.Item()
-            item.cart = self.cart
-            item.product = product
-            item.unit_price = unit_price
-            item.quantity = quantity
-            item.save()
-        else: #ItemAlreadyExists
+        item = models.Item.objects.filter(cart=self.cart, product=product).first()
+        if item:
             item.unit_price = unit_price
             item.quantity += int(quantity)
             item.save()
+        else:
+            models.Item.objects.create(cart=self.cart, product=product, unit_price=unit_price, quantity=quantity)
 
     def remove(self, product):
-        try:
-            item = models.Item.objects.get(
-                cart=self.cart,
-                product=product,
-            )
-        except models.Item.DoesNotExist:
-            raise ItemDoesNotExist
-        else:
+        item = models.Item.objects.filter(cart=self.cart, product=product).first()
+        if item:
             item.delete()
+        else:
+            raise ItemDoesNotExist
 
     def update(self, product, quantity, unit_price=None):
-        try:
-            item = models.Item.objects.get(
-                cart=self.cart,
-                product=product,
-            )
-        except models.Item.DoesNotExist:
-            raise ItemDoesNotExist
-        else: #ItemAlreadyExists
+        item = models.Item.objects.filter(cart=self.cart, product=product).first()
+        if item:
             if quantity == 0:
                 item.delete()
             else:
                 item.unit_price = unit_price
                 item.quantity = int(quantity)
                 item.save()
+        else:
+            raise ItemDoesNotExist
 
     def count(self):
-        result = 0
-        for item in self.cart.item_set.all():
-            result += 1 * item.quantity
-        return result
+        return self.cart.item_set.all().aggregate(Sum('quantity')).get('quantity__sum', 0)
 
     def summary(self):
-        result = 0
-        for item in self.cart.item_set.all():
-            result += item.total_price
-        return result
+        return self.cart.item_set.all().aggregate(total=Sum(F('quantity')*F('unit_price'))).get('total', 0)
 
     def clear(self):
-        for item in self.cart.item_set.all():
-            item.delete()
+        self.cart.item_set.all().delete()
 
     def is_empty(self):
         return self.count() == 0
@@ -99,10 +77,10 @@ class Cart:
     def cart_serializable(self):
         representation = {}
         for item in self.cart.item_set.all():
-            itemID = str(item.object_id)
-            itemToDict = {
+            item_id = str(item.object_id)
+            item_dict = {
                 'total_price': item.total_price,
                 'quantity': item.quantity
             }
-            representation[itemID] = itemToDict
+            representation[item_id] = item_dict
         return representation
